Migrate alert to typescript
