+++ frontend/package.json
@@ -30,13 +30,16 @@
     "typeface-roboto": "^0.0.54"
   },
   "devDependencies": {
+    "@types/classnames": "^2.2.7",
+    "@types/react-redux": "^7.0.9",
     "babel-eslint": "^10.0.1",
     "eslint": "^5.16.0",
     "eslint-config-airbnb": "^17.1.0",
     "eslint-plugin-import": "^2.17.2",
     "eslint-plugin-jsx-a11y": "^6.2.1",
     "eslint-plugin-react": "^7.13.0",
-    "prettier": "^1.17.1"
+    "prettier": "^1.17.1",
+    "typescript": "^3.4.5"
   },
   "eslintConfig": {
     "extends": "react-app"
