+++ frontend/src/components/Alert/SnackbarContentWrapper.styles.ts
@@ -1,0 +1,31 @@
+import { StyleRulesCallback } from '@material-ui/core/styles';
+import amber from '@material-ui/core/colors/amber';
+import green from '@material-ui/core/colors/green';
+
+const styles: StyleRulesCallback = theme => ({
+  success: {
+    backgroundColor: green[600],
+  },
+  error: {
+    backgroundColor: theme.palette.error.dark,
+  },
+  info: {
+    backgroundColor: theme.palette.primary.dark,
+  },
+  warning: {
+    backgroundColor: amber[700],
+  },
+  icon: {
+    fontSize: 20,
+  },
+  iconVariant: {
+    opacity: 0.9,
+    marginRight: theme.spacing.unit,
+  },
+  message: {
+    display: 'flex',
+    alignItems: 'center',
+  },
+});
+
+export default styles;
